refactors inspector into indiidual workspace view directories
