+++ web-local/src/lib/components/workspace/core/Inspector.svelte
