+++ web-local/src/lib/components/workspace/model/ModelWorkspace.svelte
@@ -2,7 +2,7 @@
   import { EntityType } from "@rilldata/web-local/common/data-modeler-state-service/entity-state-service/EntityStateService";
   import { dataModelerService } from "@rilldata/web-local/lib/application-state-stores/application-store";
   import WorkspaceContainer from "../core/WorkspaceContainer.svelte";
-  import ModelInspector from "../inspector/model/ModelInspector.svelte";
+  import ModelInspector from "./inspector/ModelInspector.svelte";
   import ModelBody from "./ModelBody.svelte";
   export let modelID;
