+++ web-local/src/lib/components/workspace/model/inspector/ModelInspector.svelte
