+++ web-local/src/lib/components/workspace/model/inspector/ModelInspectorModelProfile.svelte
