+++ web-local/src/lib/components/workspace/model/inspector/header/ModelInspectorHeader.svelte
