+++ web-local/src/lib/components/workspace/source/SourceInspector.svelte
